feat用策略树完成多线程异步数据加载并完成测试

+++ group-buy-market-ywz-infrastructure/src/main/java/com/ywz/infrastructure/dao/po/GroupBuyDiscountPO.java
@@ -42,7 +42,7 @@ public class GroupBuyDiscountPO {
     /**
      * 折扣类型（0:base、1:tag）
      */
-    private Integer discountType;
+    private Byte discountType;
 
     /**
      * 营销优惠计划（ZJ:直减、MJ:满减、N元购）
